refactor(flydsl): rename blockwise FP8 quant module
Remove the PoC suffix now that the FlyDSL quantization kernels are used by the production blockwise GEMM path.
